fix(cli): share dashboard rebuild helpers and bind standalone next to HOST
